requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
crispy-bootstrap5==2025.4
cryptography==44.0.2
dj-rest-auth==7.0.1
Django==5.2
django-allauth==65.7.0
django-crispy-forms==2.3
djangorestframework==3.16.0
idna==3.10
pycparser==2.22
PyJWT==2.10.1
requests==2.32.3
sqlparse==0.5.3
urllib3==2.3.0
